fix: address review feedback on unequal-split messaging
Reworks the unequal-split wording and guard per the inline review.
Lint, type-check, 141 tests, and a production build verified.
